requirements.txt
blinker==1.8.2
cbor2==5.6.4
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
docker==7.1.0
Flask==3.0.3
Flask-Cors==4.0.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
keyboard==0.13.5
luma==2.3.18
luma.core==2.4.2
luma.oled==3.13.0
MarkupSafe==2.1.5
numpy==2.0.1
pcf8574==0.1.3
pillow==10.4.0
pycparser==2.21
pyFluidSynth==1.3.3
pyftdi==0.55.4
pygame==2.6.0
pyparsing==3.0.9
pyserial==3.5
python-dateutil==2.8.0
pyusb==1.2.1
requests==2.32.3
RPi.GPIO==0.7.1
six==1.16.0
smbus-cffi==0.5.1
smbus2==0.4.3
spidev==3.6
urllib3==2.2.2
Werkzeug==3.0.3
